// web/src/app/api/generated/src/models/EditorialCatalogItem.ts
export interface EditorialCatalogItem {
  /**
   *
   * @type {Array<string>}
   * @memberof EditorialCatalogItem
   */
  availableActions?: Array<string>;
  /**
   *
   * @type {string}
   * @memberof EditorialCatalogItem
   */
  canonicalName?: string;
  /**
   *
   * @type {EditorialCatalogItemEnvironmentEnum}
   * @memberof EditorialCatalogItem
   */
  environment?: EditorialCatalogItemEnvironmentEnum;
  /**
   *
   * @type {string}
   * @memberof EditorialCatalogItem
   */
  exerciseId?: string;
  /**
   *
   * @type {number}
   * @memberof EditorialCatalogItem
   */
  expectedVersion?: number;
  /**
   *
   * @type {EditorialCatalogItemStatusEnum}
   * @memberof EditorialCatalogItem
   */
  status?: EditorialCatalogItemStatusEnum;
  /**
   *
   * @type {EditorialCatalogItemTechnicalLevelEnum}
   * @memberof EditorialCatalogItem
   */
  technicalLevel?: EditorialCatalogItemTechnicalLevelEnum;
  /**
   *
   * @type {string}
   * @memberof EditorialCatalogItem
   */
  versionId?: string;
  /**
   *
   * @type {number}
   * @memberof EditorialCatalogItem
   */
  versionNumber?: number;
}

/**
 * @export
 */
export const EditorialCatalogItemEnvironmentEnum = {
  Home: 'HOME',
  Gym: 'GYM',
  Outdoor: 'OUTDOOR',
  Clinic: 'CLINIC',
  Any: 'ANY',
} as const;
export type EditorialCatalogItemEnvironmentEnum =
  (typeof EditorialCatalogItemEnvironmentEnum)[keyof typeof EditorialCatalogItemEnvironmentEnum];

/**
 * @export
 */
export const EditorialCatalogItemStatusEnum = {
  Draft: 'DRAFT',
  InReview: 'IN_REVIEW',
  ChangesRequested: 'CHANGES_REQUESTED',
  Approved: 'APPROVED',
  Published: 'PUBLISHED',
  Withdrawn: 'WITHDRAWN',
} as const;
export type EditorialCatalogItemStatusEnum =
  (typeof EditorialCatalogItemStatusEnum)[keyof typeof EditorialCatalogItemStatusEnum];

/**
 * @export
 */
export const EditorialCatalogItemTechnicalLevelEnum = {
  Foundational: 'FOUNDATIONAL',
  Intermediate: 'INTERMEDIATE',
  Advanced: 'ADVANCED',
} as const;
export type EditorialCatalogItemTechnicalLevelEnum =
  (typeof EditorialCatalogItemTechnicalLevelEnum)[keyof typeof EditorialCatalogItemTechnicalLevelEnum];